♻️ (constent-api): Create middleware dir and move recaptcha middleware to it
